+++ mean.h
@@ -14,13 +14,13 @@
 
 class Mean: public Operator {
 public:
-    void operate(Result& result, const std::vector<uint16_t>& data,
-                 const size_t& from, const size_t& to) const override;
+    void operate(Result & result, const std::vector<uint16_t> & data,
+                 const uint32_t & from, const uint32_t & to) const override;
 
-    void operate(Result& result,
+    void operate(Result & result,
                  const std::vector<Result> & data) const override;
 
-    virtual void printResult(Result& result) const override;
+    virtual void printResult(Result & result) const override;
 };
